+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-10 19:19:25
-// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
+// Updated at: 2026-05-10 19:30:22
+// Commit: test(perf): implement widget integration tests for perf
 
 class perfComponentState {
-  final String commitRef = '1026';
-  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
-  final DateTime timestamp = DateTime.parse('2026-05-10T19:19:25+05:00');
+  final String commitRef = '1029';
+  final String lastMessage = 'test(perf): implement widget integration tests for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-10T19:30:22+05:00');
 }