+++ CHANGELOG.md
@@ -1030,3 +1030,4 @@
 - [2026-05-10 19:19] perf(perf): apply RepaintBoundary on animated perf widgets
 - [2026-05-10 19:27] perf(forms): apply RepaintBoundary on animated forms widgets
 - [2026-05-10 19:28] style(storage): adjust design tokens, typography, and contrast in storage
+- [2026-05-10 19:30] test(perf): implement widget integration tests for perf